requirements.txt
pytest
pytest-playwright
pytest-html
requests
python-dotenv
